adjust logic
