--- src/container/MentorProfile/components/MentorProfileSection/types.ts
@@ -1,6 +1,0 @@
-import { MentorT } from '~/api/mentor/types'
-
-export interface Props {
-    mentorId: string
-    mentorInfo: MentorT
-}
